// src/app/enums/storage-keys.enum.ts
/** Keys used with @capacitor/preferences (SharedPreferences / UserDefaults). */
export enum StorageKey {
  WatchSettings = 'bg-geo.watch-settings',
  UploadConfig = 'bg-geo.upload-config',
  TrackHistory = 'bg-geo.track-history',
  SessionStartedAt = 'bg-geo.session-started-at',
  SessionProgress = 'bg-geo.session-progress',
}